+++ src/css/Available.css
@@ -1,0 +1,17 @@
+.available {
+  margin: 0;
+  padding: 0;
+  border: 0;
+  background: none;
+  display: flex;
+  margin-left: auto;
+  align-items: center;
+  justify-content: center;
+  cursor: pointer;
+}
+
+.available__icon {
+  margin-right: 12px;
+  width: 16px;
+  height: 12px;
+}
