+++ src/css/Sort.css
@@ -1,29 +1,4 @@
-.availavble {
-  margin: 0;
-  padding: 0;
-  border: 0;
-  background: none;
-  display: flex;
-  margin-left: auto;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-}
-
-.availavble__icon {
-  margin-right: 12px;
-  width: 16px;
-  height: 12px;
-}
-
 .sort {
-  display: flex;
-  margin-bottom: 80px;
-  align-items: center;
-  justify-content: center;
-}
-
-.sort-by {
   display: flex;
 }
